File struct refactoring
